cleaned up server context code
